+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Changelog
 
 ## vNext
 
+### Features
+
+- Add Autocomplete clear feature [PR 49](https://github.com/input-output-hk/react-polymorph/pull/49)
+
 ## 0.6.4
 
 ### Fixes